Fix a bug in index.html
Elm was taking control of a text node, not the div node that we wanted to give it. As a result, there was a superfluous div left floating at the bottom of the the HTML.

To fix this problem, we had to change document.body.firstChild to document.body.firstElementChild; the latter skips text nodes.