Simplify Flask's configuration
Not bothering with a settings file—we don't really have any settings at the moment, and keeping around the settings file—not controlled by git—would make deployment and CI too complicated.

If, in the future, we need to have different settings for different environments, we might try to use environment variables for that instead. It's easier to set some environment variables than to write a whole config file every time we deploy. (Environment variables would play well with GitHub's Secrets, for example.)