Add minification for the JavaScript and CSS
Using UglifyJS 3 and clean-css.

Honestly, this step probably isn't that important--the uniminified
assets are only a couple hundred kilobytes anyway. Still, I'd rather
have the minification than not. To some extent, I added minification to
protect against future growth in the codebase and assets, but mostly, I
just added minification as a learning experience. It's not strictly
necessary, but it's useful to see how these tools would be used in a
real project!