Only use requirements.txt to manage dependencies
Previously, we were also using setup.py, which was a little too
Don't-Repeat-Yourself-ish.

For anyone curious:

Part of the reason we weren't using requirements.txt to manage
dependencies--beyond a general lack of experience and understanding--was
that I didn't know how to make different requirement.txt files for
development and production. However, I just learned that you can link to
one requirements.txt from another, which gives an easy solution!

See this wonderful person's answer: https://stackoverflow.com/a/20720019

There's one lingering problem with this approach, which is that

1) We can no longer user `pip freeze` to create requirements.txt, since
we have to sort the packages into categories manually, so
2) we really can only specify top-level dependencies--specifying
transitive dependencies would be too much administrative work--and so
3) We can't lock our transitive dependencies to specific versions,
which potentially makes deployments unreliable.

That's a serious issue, but I'm willing to put up it. The risk isn't
that high, and the reward is worth it!