Remove the testing dependency on pytest-runner.
pytest-runner is deprecated because it uses setuptools's
`test_requires` option, which is a little bit magic. (We stopped using
`test_requires` in our project a while ago, but we kept pytest-runner
around.)

To replace pytest-runner, we have a little class that extends
setuptools.Command and wraps the `pytest.main()` function.