Test that the input is cleared at the right times
This test is failing, of course, since we haven't implemented that
functionality yet.

These tests actually uncovered a bug where, if a GET request fails, the
entire page will be replaced with an error message. This is the desired
behavior for the *very first* GET request, but it's the wrong behavior
for every GET request after that.