No need to set Cache-Control to private or public
I set Cache-Control: private as the default in nginx.prod.conf because I
was worried that--in the event I ever add authentication/authorization
to this project--it would be dangerous to put information like someone's
watchlist in a public cache. However, it looks like HTTP has sensible
defaults in this case: anything with an Authorization header won't be
cached.

Source:
https://sookocheff.com/post/api/effective-caching/#handling-authentication