For caching, say is an even 30 days
I want certain assets to be cached for about a month, but `Cache-Control:
max-age` is expressed in seconds.

Previously, we were using the number of seconds in an average month
(ie, the number of seconds in an average year divided by 12).

Since I don't care about exactness here--I don't need to track the
Gregorian calendar or anything ;)--we can just round and say that we'll
cache assets for 30 days.